ESDEV-3462 Extract abstract test base class and little improvements.
Extract common base test class for easier classes, added comments and renamed parameters for better readability.
